Feature/validations and exception handling

+++ src/main/java/com/e3/test/service/EmployeeService.java
@@ -4,6 +4,8 @@ import com.e3.test.dto.EmployeeRequestDto;
 import com.e3.test.dto.EmployeeSearchDto;
 import com.e3.test.entity.Company;
 import com.e3.test.entity.Employee;
+import com.e3.test.exception.BusinessException;
+import com.e3.test.exception.ValidationException;
 import com.e3.test.repository.CompanyRepository;
 import com.e3.test.repository.EmployeeRepository;
 import org.springframework.data.domain.Sort;
@@ -27,14 +29,29 @@ public class EmployeeService {
     }
 
     public Employee getEmployeeById(Long employeeId) {
-        return employeeRepository.findOne(employeeId);
+        Employee employee = employeeRepository.findOne(employeeId);
+
+        if (employee == null) {
+            throw new BusinessException("ERR_003");
+        }
+
+        return employee;
     }
 
-    public Boolean deleteEmployeeById(Long employeeId) {
-        employeeRepository.delete(employeeId);
+    public void deleteEmployeeById(Long employeeId) {
+        Employee employee = employeeRepository.findOne(employeeId);
 
-        // TODO implement businessexception
-        return Boolean.TRUE;
+        if (employee == null) {
+            throw new BusinessException("ERR_003");
+        }
+
+        try {
+            employeeRepository.delete(employeeId);
+        }
+        catch (Exception ex) {
+            throw new BusinessException("ERR_004");
+        }
+
     }
 
     public List<Employee> searchEmployees(EmployeeSearchDto searchDto) {
@@ -68,6 +85,12 @@ public class EmployeeService {
     }
 
     public Employee saveEmployee(EmployeeRequestDto requestDto) {
+
+        if (requestDto.getId() == null &&
+                employeeRepository.isEmployeeAlreadyRegistered(requestDto.getFirstName(), requestDto.getLastName()) != 0) {
+            throw new ValidationException("ERR_006");
+        }
+
         Company company = companyRepository.findOne(requestDto.getCompanyId());
 
         if (company != null) {
@@ -77,8 +100,7 @@ public class EmployeeService {
             return employeeRepository.save(employee);
         }
         else {
-            // TODO throw businessexception
-            return null;
+            throw new BusinessException("ERR_002");
         }
     }
 }
